Fix sonar issues, Add pkg selector on copy, bumped version minor 1
