+++ cmd/aem/commandPackageRebuild.go
@@ -34,7 +34,8 @@ func (p *commandPackageRebuild) Execute(args []string) {
 	if len(p.Package) > 0 {
 		pkgs = p.utility.pkgsFromString(fromInstance, p.Package)
 	} else {
-		pkgs = p.utility.pkgPicker(fromInstance)
+		pkgPicker := newPackagePicker()
+		pkgs = pkgPicker.picker(fromInstance)
 	}
 
 	for _, pkg := range pkgs {
